rename everything from 'bigsky' or 'medsky' to 'relay'
