+++ cmd/relay/events/persist.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/bluesky-social/indigo/cmd/medsky/models"
+	"github.com/bluesky-social/indigo/cmd/relay/models"
 )
 
 // Note that this interface looks generic, but some persisters might only work with RepoAppend or LabelLabels
